requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.2
numpy>=1.22.0
scipy>=1.7.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.6.0
pillow>=9.0.0

# Training utilities
tqdm>=4.64.0
tensorboard>=2.11.0

# Optional: for tensor operations
einops>=0.6.0
